Remove separate flags in favor of -redis-url
All of the flags that were remove (address, db, and password) can all be
specified in the URL. The URL can also specify protocol (plaintext vs
tls) and connecting to Redis on a unix socket. Having both separate
options and the URL could be confusing since it's not clear which has
precedence (currently the url).